@@ -32,14 +32,15 @@ VERBOSE_TMPL = '''
       placeholder: |
         [debug] Command-line config: ['-vU', 'https://www.youtube.com/watch?v=BaW_jenozKc']
         [debug] Encodings: locale cp65001, fs utf-8, pref cp65001, out utf-8, error utf-8, screen utf-8
-        [debug] yt-dlp version nightly@... from yt-dlp/yt-dlp [b634ba742] (win_exe)
-        [debug] Python 3.8.10 (CPython 64bit) - Windows-10-10.0.22000-SP0
-        [debug] exe versions: ffmpeg N-106550-g072101bd52-20220410 (fdk,setts), ffprobe N-106624-g391ce570c8-20220415, phantomjs 2.1.1
-        [debug] Optional libraries: Cryptodome-3.15.0, brotli-1.0.9, certifi-2022.06.15, mutagen-1.45.1, sqlite3-2.6.0, websockets-10.3
+        [debug] yt-dlp version nightly@... from yt-dlp/yt-dlp-nightly-builds [1a176d874] (win_exe)
+        [debug] Python 3.10.11 (CPython AMD64 64bit) - Windows-10-10.0.20348-SP0 (OpenSSL 1.1.1t  7 Feb 2023)
+        [debug] exe versions: ffmpeg 7.0.2 (setts), ffprobe 7.0.2
+        [debug] Optional libraries: Cryptodome-3.21.0, brotli-1.1.0, certifi-2024.08.30, curl_cffi-0.5.10, mutagen-1.47.0, requests-2.32.3, sqlite3-3.40.1, urllib3-2.2.3, websockets-13.1
         [debug] Proxy map: {}
-        [debug] Request Handlers: urllib, requests
-        [debug] Loaded 1893 extractors
-        [debug] Fetching release info: https://api.github.com/repos/yt-dlp/yt-dlp-nightly-builds/releases/latest
+        [debug] Request Handlers: urllib, requests, websockets, curl_cffi
+        [debug] Loaded 1838 extractors
+        [debug] Fetching release info: https://api.github.com/repos/yt-dlp/yt-dlp/releases/latest
+        Latest version: nightly@... from yt-dlp/yt-dlp-nightly-builds
         yt-dlp is up to date (nightly@... from yt-dlp/yt-dlp-nightly-builds)
         [youtube] Extracting URL: https://www.youtube.com/watch?v=BaW_jenozKc
         <more lines>
