@@ -82,16 +82,32 @@ TEST_LOCKFILE_V1 = rf'''{TEST_LOCKFILE_COMMENT}
 lock 2022.08.18.36 .+ Python 3\.6
 lock 2023.11.16 (?!win_x86_exe).+ Python 3\.7
 lock 2023.11.16 win_x86_exe .+ Windows-(?:Vista|2008Server)
+lock 2024.10.22 py2exe .+
+lock 2024.10.22 linux_(?:armv7l|aarch64)_exe .+-glibc2\.(?:[12]?\d|30)\b
+lock 2024.10.22 (?!\w+_exe).+ Python 3\.8
+lock 2024.10.22 win(?:_x86)?_exe Python 3\.[78].+ Windows-(?:7-|2008ServerR2)
 '''
 
 TEST_LOCKFILE_V2_TMPL = r'''%s
 lockV2 yt-dlp/yt-dlp 2022.08.18.36 .+ Python 3\.6
 lockV2 yt-dlp/yt-dlp 2023.11.16 (?!win_x86_exe).+ Python 3\.7
 lockV2 yt-dlp/yt-dlp 2023.11.16 win_x86_exe .+ Windows-(?:Vista|2008Server)
+lockV2 yt-dlp/yt-dlp 2024.10.22 py2exe .+
+lockV2 yt-dlp/yt-dlp 2024.10.22 linux_(?:armv7l|aarch64)_exe .+-glibc2\.(?:[12]?\d|30)\b
+lockV2 yt-dlp/yt-dlp 2024.10.22 (?!\w+_exe).+ Python 3\.8
+lockV2 yt-dlp/yt-dlp 2024.10.22 win(?:_x86)?_exe Python 3\.[78].+ Windows-(?:7-|2008ServerR2)
 lockV2 yt-dlp/yt-dlp-nightly-builds 2023.11.15.232826 (?!win_x86_exe).+ Python 3\.7
 lockV2 yt-dlp/yt-dlp-nightly-builds 2023.11.15.232826 win_x86_exe .+ Windows-(?:Vista|2008Server)
+lockV2 yt-dlp/yt-dlp-nightly-builds 2024.10.22.051025 py2exe .+
+lockV2 yt-dlp/yt-dlp-nightly-builds 2024.10.22.051025 linux_(?:armv7l|aarch64)_exe .+-glibc2\.(?:[12]?\d|30)\b
+lockV2 yt-dlp/yt-dlp-nightly-builds 2024.10.22.051025 (?!\w+_exe).+ Python 3\.8
+lockV2 yt-dlp/yt-dlp-nightly-builds 2024.10.22.051025 win(?:_x86)?_exe Python 3\.[78].+ Windows-(?:7-|2008ServerR2)
 lockV2 yt-dlp/yt-dlp-master-builds 2023.11.15.232812 (?!win_x86_exe).+ Python 3\.7
 lockV2 yt-dlp/yt-dlp-master-builds 2023.11.15.232812 win_x86_exe .+ Windows-(?:Vista|2008Server)
+lockV2 yt-dlp/yt-dlp-master-builds 2024.10.22.045052 py2exe .+
+lockV2 yt-dlp/yt-dlp-master-builds 2024.10.22.060347 linux_(?:armv7l|aarch64)_exe .+-glibc2\.(?:[12]?\d|30)\b
+lockV2 yt-dlp/yt-dlp-master-builds 2024.10.22.060347 (?!\w+_exe).+ Python 3\.8
+lockV2 yt-dlp/yt-dlp-master-builds 2024.10.22.060347 win(?:_x86)?_exe Python 3\.[78].+ Windows-(?:7-|2008ServerR2)
 '''
 
 TEST_LOCKFILE_V2 = TEST_LOCKFILE_V2_TMPL % TEST_LOCKFILE_COMMENT
@@ -145,43 +161,76 @@ class TestUpdate(unittest.TestCase):
         for lockfile in (TEST_LOCKFILE_V1, TEST_LOCKFILE_V2, TEST_LOCKFILE_ACTUAL, TEST_LOCKFILE_FORK):
             # Normal operation
             test(lockfile, 'zip Python 3.12.0', '2023.12.31', '2023.12.31')
-            test(lockfile, 'zip stable Python 3.12.0', '2023.12.31', '2023.12.31', exact=True)
-            # Python 3.6 --update should update only to its lock
+            test(lockfile, 'zip Python 3.12.0', '2023.12.31', '2023.12.31', exact=True)
+            # py2exe should never update beyond 2024.10.22
+            test(lockfile, 'py2exe Python 3.8', '2025.01.01', '2024.10.22')
+            test(lockfile, 'py2exe Python 3.8', '2025.01.01', None, exact=True)
+            # Python 3.6 --update should update only to the py3.6 lock
             test(lockfile, 'zip Python 3.6.0', '2023.11.16', '2022.08.18.36')
-            # --update-to an exact version later than the lock should return None
-            test(lockfile, 'zip stable Python 3.6.0', '2023.11.16', None, exact=True)
-            # Python 3.7 should be able to update to its lock
+            # Python 3.6 --update-to an exact version later than the py3.6 lock should return None
+            test(lockfile, 'zip Python 3.6.0', '2023.11.16', None, exact=True)
+            # Python 3.7 should be able to update to the py3.7 lock
             test(lockfile, 'zip Python 3.7.0', '2023.11.16', '2023.11.16')
-            test(lockfile, 'zip stable Python 3.7.1', '2023.11.16', '2023.11.16', exact=True)
-            # Non-win_x86_exe builds on py3.7 must be locked
+            test(lockfile, 'zip Python 3.7.1', '2023.11.16', '2023.11.16', exact=True)
+            # Non-win_x86_exe builds on py3.7 must be locked at py3.7 lock
             test(lockfile, 'zip Python 3.7.1', '2023.12.31', '2023.11.16')
-            test(lockfile, 'zip stable Python 3.7.1', '2023.12.31', None, exact=True)
-            test(  # Windows Vista w/ win_x86_exe must be locked
-                lockfile, 'win_x86_exe stable Python 3.7.9 (CPython x86 32bit) - Windows-Vista-6.0.6003-SP2',
+            test(lockfile, 'zip Python 3.7.1', '2023.12.31', None, exact=True)
+            # Python 3.8 should only update to the py3.8 lock
+            test(lockfile, 'zip Python 3.8.10', '2025.01.01', '2024.10.22')
+            test(lockfile, 'zip Python 3.8.110', '2025.01.01', None, exact=True)
+            test(  # Windows Vista w/ win_x86_exe must be locked at Vista lock
+                lockfile, 'win_x86_exe Python 3.7.9 (CPython x86 32bit) - Windows-Vista-6.0.6003-SP2',
                 '2023.12.31', '2023.11.16')
-            test(  # Windows 2008Server w/ win_x86_exe must be locked
+            test(  # Windows 2008Server w/ win_x86_exe must be locked at Vista lock
                 lockfile, 'win_x86_exe Python 3.7.9 (CPython x86 32bit) - Windows-2008Server',
                 '2023.12.31', None, exact=True)
-            test(  # Windows 7 w/ win_x86_exe py3.7 build should be able to update beyond lock
-                lockfile, 'win_x86_exe stable Python 3.7.9 (CPython x86 32bit) - Windows-7-6.1.7601-SP1',
-                '2023.12.31', '2023.12.31')
-            test(  # Windows 8.1 w/ '2008Server' in platform string should be able to update beyond lock
+            test(  # Windows 7 w/ win_x86_exe py3.7 build should be able to update beyond py3.7 lock
+                lockfile, 'win_x86_exe Python 3.7.9 (CPython x86 32bit) - Windows-7-6.1.7601-SP1',
+                '2023.12.31', '2023.12.31', exact=True)
+            test(  # Windows 7 win_x86_exe should only update to Win7 lock
+                lockfile, 'win_x86_exe Python 3.7.9 (CPython x86 32bit) - Windows-7-6.1.7601-SP1',
+                '2025.01.01', '2024.10.22')
+            test(  # Windows 2008ServerR2 win_exe should only update to Win7 lock
+                lockfile, 'win_exe Python 3.8.10 (CPython x86 32bit) - Windows-2008ServerR2',
+                '2025.12.31', '2024.10.22')
+            test(  # Windows 8.1 w/ '2008Server' in platform string should be able to update beyond py3.7 lock
                 lockfile, 'win_x86_exe Python 3.7.9 (CPython x86 32bit) - Windows-post2008Server-6.2.9200',
                 '2023.12.31', '2023.12.31', exact=True)
+            test(  # win_exe built w/Python 3.8 on Windows>=8 should be able to update beyond py3.8 lock
+                lockfile, 'win_exe Python 3.8.10 (CPython AMD64 64bit) - Windows-10-10.0.20348-SP0',
+                '2025.01.01', '2025.01.01', exact=True)
+            test(  # linux_armv7l_exe w/glibc2.7 should only update to glibc<2.31 lock
+                lockfile, 'linux_armv7l_exe Python 3.8.0 (CPython armv7l 32bit) - Linux-6.5.0-1025-azure-armv7l-with-glibc2.7',
+                '2025.01.01', '2024.10.22')
+            test(  # linux_armv7l_exe w/Python 3.8 and glibc>=2.31 should be able to update beyond py3.8 and glibc<2.31 locks
+                lockfile, 'linux_armv7l_exe Python 3.8.0 (CPython armv7l 32bit) - Linux-6.5.0-1025-azure-armv7l-with-glibc2.31',
+                '2025.01.01', '2025.01.01')
+            test(  # linux_armv7l_exe w/glibc2.30 should only update to glibc<2.31 lock
+                lockfile, 'linux_armv7l_exe Python 3.8.0 (CPython armv7l 64bit) - Linux-6.5.0-1025-azure-aarch64-with-glibc2.30 (OpenSSL',
+                '2025.01.01', '2024.10.22')
+            test(  # linux_aarch64_exe w/glibc2.17 should only update to glibc<2.31 lock
+                lockfile, 'linux_aarch64_exe Python 3.8.0 (CPython aarch64 64bit) - Linux-6.5.0-1025-azure-aarch64-with-glibc2.17',
+                '2025.01.01', '2024.10.22')
+            test(  # linux_aarch64_exe w/glibc2.40 and glibc>=2.31 should be able to update beyond py3.8 and glibc<2.31 locks
+                lockfile, 'linux_aarch64_exe Python 3.8.0 (CPython aarch64 64bit) - Linux-6.5.0-1025-azure-aarch64-with-glibc2.40',
+                '2025.01.01', '2025.01.01')
+            test(  # linux_aarch64_exe w/glibc2.3 should only update to glibc<2.31 lock
+                lockfile, 'linux_aarch64_exe Python 3.8.0 (CPython aarch64 64bit) - Linux-6.5.0-1025-azure-aarch64-with-glibc2.3 (OpenSSL',
+                '2025.01.01', '2024.10.22')
 
         # Forks can block updates to non-numeric tags rather than lock
         test(TEST_LOCKFILE_FORK, 'zip Python 3.6.3', 'pr0000', None, repo='fork/yt-dlp')
-        test(TEST_LOCKFILE_FORK, 'zip stable Python 3.7.4', 'pr0000', 'pr0000', repo='fork/yt-dlp')
-        test(TEST_LOCKFILE_FORK, 'zip stable Python 3.7.4', 'pr1234', None, repo='fork/yt-dlp')
+        test(TEST_LOCKFILE_FORK, 'zip Python 3.7.4', 'pr0000', 'pr0000', repo='fork/yt-dlp')
+        test(TEST_LOCKFILE_FORK, 'zip Python 3.7.4', 'pr1234', None, repo='fork/yt-dlp')
         test(TEST_LOCKFILE_FORK, 'zip Python 3.8.1', 'pr1234', 'pr1234', repo='fork/yt-dlp', exact=True)
         test(
-            TEST_LOCKFILE_FORK, 'win_x86_exe stable Python 3.7.9 (CPython x86 32bit) - Windows-Vista-6.0.6003-SP2',
+            TEST_LOCKFILE_FORK, 'win_x86_exe Python 3.7.9 (CPython x86 32bit) - Windows-Vista-6.0.6003-SP2',
             'pr1234', None, repo='fork/yt-dlp')
         test(
-            TEST_LOCKFILE_FORK, 'win_x86_exe stable Python 3.7.9 (CPython x86 32bit) - Windows-7-6.1.7601-SP1',
+            TEST_LOCKFILE_FORK, 'win_x86_exe Python 3.7.9 (CPython x86 32bit) - Windows-7-6.1.7601-SP1',
             '2023.12.31', '2023.12.31', repo='fork/yt-dlp')
         test(TEST_LOCKFILE_FORK, 'zip Python 3.11.2', 'pr9999', None, repo='fork/yt-dlp', exact=True)
-        test(TEST_LOCKFILE_FORK, 'zip stable Python 3.12.0', 'pr9999', 'pr9999', repo='fork/yt-dlp')
+        test(TEST_LOCKFILE_FORK, 'zip Python 3.12.0', 'pr9999', 'pr9999', repo='fork/yt-dlp')
 
     def test_query_update(self):
         ydl = FakeYDL()
