@@ -1,5 +1,6 @@
+import functools
+
 from .common import InfoExtractor
-from ..compat import functools
 from ..utils import (
     int_or_none,
     parse_duration,
