@@ -1,8 +1,8 @@
 # Autogenerated by devscripts/update-version.py
 
-__version__ = '2024.10.07'
+__version__ = '2024.10.22'
 
-RELEASE_GIT_HEAD = '1a176d874e6772cd898ce507379ea388e96ee3f7'
+RELEASE_GIT_HEAD = '67adeb7bab00662ba55d473e405b301abb42fe61'
 
 VARIANT = None
 
@@ -12,4 +12,4 @@ CHANNEL = 'stable'
 
 ORIGIN = 'yt-dlp/yt-dlp'
 
-_pkg_version = '2024.10.07'
+_pkg_version = '2024.10.22'
